Update indexedList.lua
Added many useful functions available to all derivatives of these classes.

Add a metatable to `IndexedList` and `IndexedEntry`, to filter out fields starting with `_` when iterating it with `table.pairs`.
Add a table `_default` to `IndexedEntry`.
`IndexedEntry` starts out in an "unlocked" state. In its unlocked state all indexing of the table will be redirected to `_default`.
`IndexedEntry:lock()` "locks" the table, makes a deep copy of `_default` into itself, and stops redirecting to `_default`. This can be used to create a definition, lock it, and then be able to fall back to the previous state via `_default`.