Create extension of UiScrollArea and UiScrollAreaH
Extensions of classes in the mod loader.
Notable differences:

Ui elements inside the scrollarea will only be clipped in the direction of the scroll.
Supports scroll areas inside scroll areas, only scrolling the topmost element if several overlap.

New functions
scrollTo: scrolls to the target coordinate in the inner area.
scrollToContain: scrolls the shortest amount needed to contain an area in the inner area.
hasScrollbar: helps not drawing the scrollbar if its size is 0.