Create Drag Drop ui classes
UiDragObject:
A ui class which can be dragged and dropped on ui elements.
It will call its `onDropTarget-` functions when its `dragTargetType` is equal to the target's `dropTargetType`.

UiDropTarget:
An element meant to be dropped on by a `UiDragObject`. If its `dropTargetType` is the same as a `UiDragObject`'s `dragObjectType`, its `onDragObject-` functions will be called.

UiDragSource:
A class which serves as a source of `UiDragObject`'s. The drag object is set when creating the ui instance.
When the element is dragged, its drag object's `onDragSourceGrabbed` function is called.