Create UiEditBox.lua
A ui class used for ui elements which can change based on outside input. Built to work in tandem with `UiEditorButton`.
`onRecieve` and `onSend` can be overridden with functionality on the respective events.
`recieve` and `send` can be called, which if no argument is given, will call `onRecieve` and `onSend` respectively with the `easyEdit.displayedEditorButton` as the `sender` or `reciever`. `easyEdit.displayedEditorButton` in this case is the `UiEditorButton` currently hovered/selected.