Create EasyEdit configuration menu
Since EasyEdit must be initialized before mods gets a chance to properly initialize, it does not follow the normal conventions for code initialized by mods.

Due to this, an EasyEdit configuration menu will always be loaded. In which it will be possible to choose whether to initialize EasyEdit itself after restarting the game. In this way it can essentially be disabled after all.

EasyEdit config menu contains the following options:
- Toggle EasyEdit itself.
- Toggle EasyEdit debug logging.
- Toggle EasyEdit debug ui which can be used to debug ui systems.