Add ImageOffset back for tracked unit data
Tracking ImageOffset caused some issues when doing so for Mechs, since they would be dynamically set by palette systems. This in turn caused Psions and Bosses to be drawn with default colors in Easy Edit menues.

With the Mech Editor on ice, ImageOffset can once again be tracked, to correctly draw the colors for Psions and Bosses.