Fixes an Issue with randomization of Enemy Lists
Without the Fix: Steps to Recreate:
Boot the game. 
Create a New Game and start it
Record which Vek the first island contains (or more islands)
Close the game and reboot
Create a New Game and Start it
Note that the current Vek are the same as the previously recorded Vek
Repeat Any Number of Times

With the Fix: 
Repeat the same steps, but note that the Vek are properly different on the island each reboot.