Add event onGeneratingTerrain
This event is called once at the start of a mission, when terrain is generated for the mission. It uses the entry point getRainChance, which is also only checked once at the start of a mission.

For documentation:

signature of the event handler:
`void event_handler(tileset_id)`

example:
```lua
local event_handler = function(tileset_id)
    LOG("Generating terrain for tileset ", tileset)
end

easyEdit.events.onGeneratingTerrain:subscribe(event_handler)
```