Hack to prevent crash exiting world editor
How to test crash:
(Either of 1a, 1b, 1c will work)
1a. Create new profile with new name.
1b. Delete easy edit save data for a profile.
1c. Delete world.lua in the easy edit save data for a profile.

Entering and exiting the world editor after having done one of the above will crash the game.

I am not sure if this crash is introduced after the `edited` tag was introduced to easy edit entries, but it could be.
The crash happens because when world.lua is not saved due to not being different from vanilla, the cached save data is also not updated.

This hack writes to the cache when world data is updated, regardless of whether the file exists or not. So if the world data is completely vanilla, and will not be saved to file, the cache will be set to `DEFAULT_ISLAND_SLOTS`

This PR has not been tested extensively, and mostly just theory crafted.
Please do test it before merging.